requirements.txt
fastapi==0.123.4
uvicorn==0.38.0
numpy==2.2.6
pdf2image==1.17.0
pdfplumber==0.11.8
pgvector==0.4.2
opencv-python==4.12.0.88
pytesseract==0.3.13
torch==2.9.1
PyMuPDF==1.26.6
firebase_admin
sentence-transformers>=2.2.0
numpy>=1.21.0
rank-bm25>=0.2.1  # For BM25 implementation
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]

# Migrations
alembic>=1.16.0

dotenv
pydantic_settings
psycopg2
librosa
jiwer
google-genai
python-multipart

ultralytics